类型修改

+++ protocol/requests/alarm.go
@@ -3,8 +3,8 @@ package requests
 // CreateAlarmReq 创建告警配置请求
 type CreateAlarmReq struct {
 	Type   string `json:"type" validate:"required,oneof=token billing"`
-	Min    int    `json:"min" validate:"gte=0"`
-	Max    int    `json:"max" validate:"gtefield=Min"`
+	Min    int64  `json:"min" validate:"gte=0"`
+	Max    int64  `json:"max" validate:"gtefield=Min"`
 	Status *int   `json:"status"` // 使用指针区分未传值和0
 	UserId *int64 `json:"user_id"`
 }
@@ -13,8 +13,8 @@ type CreateAlarmReq struct {
 type UpdateAlarmReq struct {
 	Id     int64  `json:"id" validate:"required,gt=0"`
 	Type   string `json:"type" validate:"omitempty,oneof=token billing"`
-	Min    *int   `json:"min"`
-	Max    *int   `json:"max"`
+	Min    *int64 `json:"min"`
+	Max    *int64 `json:"max"`
 	Status *int   `json:"status"`
 	UserId *int64 `json:"user_id"`
 }
